move tanks_connected to commander
This function is generally useful for scenarios.
